+++ clang-tidy/tool/ClangTidyMain.cpp
@@ -413,6 +413,11 @@ static int clangTidyMain(int argc, const char **argv) {
   return 0;
 }
 
+// This anchor is used to force the linker to link the BugProneModule.
+extern volatile int BugProneModuleAnchorSource;
+static int LLVM_ATTRIBUTE_UNUSED BugProneModuleAnchorDestination =
+    BugProneModuleAnchorSource;
+
 // This anchor is used to force the linker to link the CERTModule.
 extern volatile int CERTModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED CERTModuleAnchorDestination =